requirements.txt
django
djangorestframework
environs
django_filters
django-redis
debug_toolbar
